requirements.txt
Django==1.6
Pillow==2.2.1
South==0.8.4
boto==2.19.0
dj-database-url==0.2.2
dj-static==0.0.5
django-registration==1.0
django-storages==1.1.8
django-toolbelt==0.0.1
fudge==1.0.3
gunicorn==18.0
humanize==0.5
psycopg2==2.5.1
python-magic==0.4.6
requests==2.2.1
simplejson==3.3.2
sorl-thumbnail==11.12
soundcloud==0.4.1
static==0.4
wsgiref==0.1.2
yolk==0.4.3
